+++ lib/ui/components.dart
@@ -17,9 +17,9 @@ export 'widgets/atoms/TabItem/tab_item.dart';
 export 'widgets/atoms/TooltipIconButton/tooltip_icon_button.dart';
 
 // Molecules - Combinations of atoms
+export 'widgets/molecules/AnalyticsSummaryCard/analytics_summary_card.dart';
 export 'widgets/molecules/HierarchicalMenuItem/hierarchical_menu_item.dart';
 export 'widgets/molecules/MenuItem/menu_item.dart';
-export 'widgets/molecules/NodeCard/node_card.dart';
 export 'widgets/molecules/Notification/notification_toast.dart';
 export 'widgets/molecules/TabMenu/tab_menu.dart';
 export 'widgets/molecules/Tooltip/tooltip.dart';