+++ doc/reference/changelog.md
@@ -7,6 +7,18 @@ et ce projet adhère au [Versioning Sémantique](https://semver.org/lang/fr/).
 
 ---
 
+## [0.3.2] - 2025-12-05
+
+### 🐛 Corrigé
+
+- **NodeCard** (#40) : Rename NodeCard to AnalyticsSummaryCard (#40)
+
+## [0.3.1] - 2025-12-05
+
+### 🐛 Corrigé
+
+- (#39) : add package properties to render icon on frontend project
+
 ## [0.3.0] - 2025-12-05
 
 ### ✨ Ajouté